Move memory and prompt calculations out of the Chat interface
